refactor(kubernetes): Require callers of getRelationships to pass data
getRelationships is inherently ineffecient, as it encourages callers to
require a lookup of data they already have.  In 2 of the 3 callers we
are either looking up the same data in a loop, or already have the data
we're looking up.

Require callers to first look up the cache data, then pass it in to
get its relationships.