refactor(kubernetes): Add Multimap-returning getRelationships
This commit only affects CacheData, in preparation for some changes
visible outside the class. Most notably, we'll update getRelationships
to return a MultiMap when we pass in a Collection of CacheData items.
This allows the caller to see the association between the input data
and the output relationships, rather than flattening it to a
collection where it's not possible to determine which of the output
CacheData items is related to each of the input ones.

In this commit, we're not taking advantage of this; in fact the
two callers have been updated to:
* Just call .values() to flatten the result
* Query the cache directly instead of using getRelationships

But in the next commit we'll push this information outward one more
level of calls, which will the allow consumers to use the information
to avoid duplicate work.

Also update getRelationshipKeys to collect to a Set. The general idea
is that we'll collect keys (String) to a Set to dedupliate but
data (CacheData) to a List as deduplication by reference is not
valuable.  (I left one TODO in a case where we are deduplicating
by reference but will fix that.)