refactor(kubernetes): Keep track of relationships when fetching
In a number of places, we have a Collection of CacheData items,
and we want to fetch all relationships of a particular kind.
For example, we might have a List of server groups and want to
load all instances.

Currently we returned a flattened collection of all the related
types, forcing the caller to then iterate through it to recompute
which instance is associated with which server group.

Greatly simplify this (and improve performance) by returning
a MultiMap from getRelationships. This means, for example, that
the caller can then look up the instances related to a given
server group id, without needing to build another map.