refactor(kubernetes): Load balancer constructor accepts server groups
The fromCacheData method to create a LoadBalancer requires all the
cache data to create server groups as well; but in some cases we
will already have constructed server groups. It will be simpler and
more performant to just pass these existing server groups to
fromCacheData.

Add a (as yet unused) fromCacheData that takes a set of
LoadBalancerServerGroups and uses those. Refactor the existing
fromCacheData to pass the LoadBalancerServerGroups it constructs
to this new constructor.