refactor(kubernetes): Minor cleanup in loadClusters
Now that only serverGroupToInstances is only used in loadClusters,
look up that data in the function rather than require the caller to
look it up.

Similarly, the construction of the stream over load balancers is
just a really complex way of getting the values of the MultiMap;
use that instead.