refactor(kubernetes): Update getServerGroup to call existing function
getServerGroup has its own logic for constructing server groups;
instead call the loadServerGroups function we just wrote.

This commit also fixes a bug; we were only loading load balancer
and instance relationships with the cache data, so we never
returned any server group managers.

The integration tests didn't catch this because they were
exercising the function only for a non-deployment-managed
replica set. Add a test to catch this case.

It's interesting that we only do this optimization of filtering
relationships when loading a single server group, whereas it
probably would be much more beneficial in the endpoints that
return many. I'll leave this for now, but may revisit.