refactor(kubernetes): Extract relationshipFilter logic
I'm honestly not sure that the benefit of creating this relationship
filter for a single server group is worth it, but we'll keep the
logic for now.  Let's at least store the cache filter as an instance
field so that we can re-use it, and abstract away the logic for
creating it.

I'm not going to include filter logic in other places as I think the
correct fix for improving performance here is just to cut down
the relationships we store in the cache to those that are actually
needed, rather than add logic for ignoring most of them.