fix(kubernetes): Stop returning multiple copies of load balancer
The issue was that we were building the load balancer once per
attached server group.

All of the refactors up to now have not fixed it (though we did
reduce the number of times it is returned from 4 to 2).

The reason at this point is because .values() on a multimap returns
items multiple times if they map to multiple keys. Let's apply a
distinct on the stream that only takes the first for each id.

We could have just done a regular distinct() because the result of
getRelationships always maps a key to the same reference (so we
don't have multiple instances of each load balancer, just multiple
pointers to the same one).  But I did not want to have to debug
this issue if that ever changes, so let's be explicit and keep
the first one we encounter with each id.