refactor(kubernetes): Push getAccount to implementations
First, getAccountName is only needed for security groups; push
that down only to the security group provider.

Second, getAccount is defined in the base class, but depends on
the abstract method getKey that is defined by implementations.
As this is the only reason we need to have the getKey method,
let's just push getAccount itself down to the implementations
so that we can just access the key directly. In fact, instead
of storing it at all, let's just use the key in the constructor
and discard it as we don't need it for anything else.

Eventually the plan will be to not read this off the key at all
in the constructor as we should already be storing the account
as part of the cache data we're using to construct this.