refactor(kubernetes): Remove manifest from view objects
This commit removes the manifest from the view object models,
instead reading the required fields from the manifest in the
constructor, then discarding the manifest.

This is an intermediate step towards not requiring the manifest
in the constructor at all, and instead using the metadata we store
on the cache object.