refactor(kubernetes): Validate account earlier and add annotations
Currently we validate a managedAccount when creating the
KubernetesV2Credentials, but by that time we've already read a
number of fields from it.  Move this validation earlier, to the
beginning of the KubernetesNamedAccountCredentials constructor
so that we can throw any errors earlier.

This will not create new exceptions for any workflow, as we were
always building the KubernetesV2Credentials at the end of this
function (and calling validate) anyway, it just moves the exception
to happen earlier.

Let's also now annotate the two places we read the accountName with
NonNull. We'll be extra careful and use Objects.requireNonNull even
though we know we've already validated with a better message earlier.
(This is so that if that validation ever changes/breaks at least we'll
still be guaranteed a non-null value as the annotation states, albeit
with a less helpful runtime exception.)