refactor(kubernetes): Remove StubKubernetesCredentials
I wrote these to make testing easier, but these implement the
generic KubernetesCredentials interface. Now that there's no V1
provider anymore, we have only one implementation of the interface
and we should eliminate it for simplicity.

Given that the interface doesn't really represent a useful level of
abstraction, using it in tests is really not that much better than
mocking, so use a mock instead.

Ideally we'll break up KubernetesV2Credentials to make testing easier
but for now let's remove the level of abstraction that no longer adds
value.