refactor(kubernetes): Clean up convertAsArtifact
convertAsArtifact is doing a lot of null checks on things that
can't be null.  If the artifact is present, it's guaranteed to
have a non-null type; likewise the getCreationTimestamp never
returns null.

Remove these checks, inline some variables, and move to an
.ifPresent to greatly simplify the function.

Finally, we were storing a relationship between the infrastructure
item and the artifact; we don't need or use these relationships at
all and treat these two areas of the cache separately. It's better
for performance on cache writes/reads, as well as simplifies the code,
if we remove these relationships. This means we don't need to
return the key from convertAsArtifact, or use it to add a relationship
in the calling code.