refactor(kubernetes): Move filtering logic to stream
Instead of exiting early from convertAsResource, just move the
filtering logic to the stream where it's called as it's clearer
there.  Also update KubernetesCachingProperties to be immutable
and only have the API that we actually use.

Finally, remove setting the application we got from the
KubernetesCachingProperties as an attribute in the cache. We don't
use this in any case, and in cases where we do want the applciation,
we should use the one on the moniker that we persist in the cache
data.