refactor(kubernetes): Clean up some logging
We check a few things on a manifest before caching and log a warning;
it's not clear that this is really useful. The check that it's not
null is always false; the check for name/namespace I guess could help
with debugging, but we're really just checking that Kubernetes is
returning a valid manifest, which seems not that useful.

At points where we actually use these, I guess we could log/fail if
they are missing, but I've never found a use for this log message in
all my time debugging Kubernetes bugs.

Similarly, logging every non-Spinnaker-owned manifest does not seem
useful; there is a flag to decide if you want to cache these or not.
If you have the flag on you want these cached, if not, you don't
reach this log statement.