refactor(kubernetes): Use Moniker from CacheData instead of re-computing
As a first step towards using the data we cache instead of re-computing
things, let's use the Moniker that we store on the CacheData. We
always store this, but are currently throwing it away and re-computing
it.

For now I'm just adding Moniker to the constructor, though my eventual
plan (once we build everything from the CacheData) will be to pass in
a strongly-typed CacheData object rather than greatly increase the
number of constructor parameters.