refactor(kubernetes): Use a Kubernetes-specific namer registry
We currently use a reasonably complex NamerRegistry to handle
two things:
(1) Keeping track of any NamingStrategy components on the classpath
and returning them by name when we call namerRegistry.getNamingStrategy.
(This happens when we construct Kubernetes credentials and have a string
representing the naming strategy.)
(2) Given an account name, asking for the naming strategy for that account.

In both cases, it's overly complicated that we have general logic for all
providers and this logic is not type-safe.  For example getNamingStrategy
returns a raw Namer, so we have to hope that is a Namer<KubernetesManifest>,
and hope is not a naming strategy.

Likewise, every time we need to look up the namer for an account, we need
to query a global registry across all providers (but first tell it that we
want a namer for a Kubernetes account, the name of the account, and that
we want to name KubernetesManifest.) While this looks somewhat typesafe,
it turns out that the only reason we pass KubernetesManifest.class to
these calls is so that withResource can cast the result for us (so
we don't need to), so we are only getting the illusion of safety.

Let's instead:
(1) Make a KubernetesNamerRegistry that keeps track of any registered
NamingStrategy<KubernetesManifest> and return them by name (while
guaranteeing typing).
(2) Once we get the namer, just store it on the credentials object
like we do for other things, instead of relying on a global registry.

I don't like having more and more things on the credentials object,
but I still think this is better than global state which is the
current state. Also, by having these things all in one place on the
object we can start to see patterns and refactor them into better
classes.

We now no longer use the general NamerRegistry at all in the
Kubernetes provider; other providers still use it, though they
may also want to just implement their own like this.