perf(kubernetes): Only load metrics if events are requested
For performance reasons we have the includeEvents flag in the
getManifest function because in most cases we don't need events.
We only need events when loading a manfiest from deck so that we
can populate them in the UI.

The same is true for metrics. We don't need metrics when checking
on the status of a deployed manifest, only when loading the UI
tab for the manifest.

Rather than add another flag, let's just consider metrics to be
events and only load them when we've requested events.