perf(kubernetes): Always do a live lookup for getManifest
Many users managing large Kubernetes deployments with Spinnaker
use the liveManifestCalls flag because it allows deployments to
complete without waiting for a cache refresh cycle. The initial
plan to remove this flag (and revert to using the cache always)
was discussed at the Kubernetes SIG and users did not want to
lose this functionality. In order to simplify the code and reduce
the configuration necessary to get a working, scalable deployment,
let's make this the default given that it has proven itself with
many large users.

This commit replaces the two kubernetes manifest providers with a
single one; it always does a live lookup for getManifest() and
always uses the cache for getClusterAndSortAscending (which was
always the case for users of liveManifestCalls because this was
never implemented in the live provider).

This will not immediately speed up deploys for users using
liveManifestCalls because we will still block on force cache refresh.
We need to do this because there are bugs downstream (ie, with
disabling/deleting older server groups) if the cache is not up to
date. The next step will be to fix this cache refresh strategy and
remove the force cache refresh stage in all cases.