fix(google): Fix assertion error on unknown target proxy
Our code to parse the target proxy type from a URL is overly strict
and throws an IllegalArgumentException if it encounter something it
doesn't recognize.

While I suppose that it might be helpful for us to know if there are
different types of proxies we don't support encountered, this is very
disruptive to users who might have these in their deployments as we
just fail.

In nearly all cases, we are just checking if the result of
getTargetProxyType is a particular type. This logic will still work
if we just return an UNKNOWN type if we don't recognize the target
proxy type, so let's do that instead.