fix(kubernetes): fix KubernetesHealthIndicatorTest
The commit that removed the StubKubernetesCredentialsFactory broke this test because it tried to create a mocked item in a `thenReturn` statement of another mock. I was finding these mocks fairly dificult to reason about when trying to make this test pass, so decided to lift most of the mocks out of static methods and into a setup block (mostly because that is what I've seen Plumpy do when creating mocks, but also because it became a bit easier to see which mocks relied on which other mocks, and which were actually unnecessary). Let me know if you think this actually just made things more confusing...