refactor(kubernetes): Minor simplifications to artifact replacer
(1) There are a lot of small functions with unclear names that
make the logic hard to follow; inline these where it makes the
logic clearer.
(2) The result of a document.read lets you iterate over the resulting
nodes, and convert them to text. This is much clearer (and will perform
better) than using an ObjectMapper for this.
(3) The Artifact passed in to replaceIfPossible is always non-null,
because it ultimately comes from the ImmutableCollection created
by filterKubernetesArtifactsByNamespaceAndAccount. That function
also filters out artifacts without a type set so we don't need to
explicitly check again (though we'll still be safe and invert the
.equals below because we can't really guarantee with an annotation
that this particular artifact has a non-null type even though it
effectively always does).