refactor(kubernetes): Move ArtifactProvider to clouddriver-kubernetes
This is only implmented/used in the kubernetes provider, and it's not
clear whether this is an abstraction useful to other providers. Let's
move it to clouddriver-kubernetes. If this becomes something useful
for other providers we can abstract common functionality to an interface
in clouddriver-core at that point.

We can probably just delete the interface as it has only one implementation,
but keeping it for now.