test(kubernetes): Add some tests on ArtifactProvider and Replacers
Let's add a few tests on the ArtifactProvider. As it depends on an
implicit contract with the caching agent, let's use the integration
test rather than mock a lot of CacheData items that look like how
the caching agent currently stores data.

In a recent PR, we converted the ArtifactReplacer tests to Java and
also improved the test coverage. While these covered the general
interface of ArtifactReplacer by using a few specific replacers,
we still don't have good test coverage of all of the replacers
defined in Replacer.java.

This is important as the functionality of these replacers is almost
entirely encoded by complex JsonPath expressions, and I'd like to
refactor these to be cleaner.

This commit adds a ReplacerTest class that goes through each defined
replacer and tries to use it to find and replace artifacts. This should
provide sufficient coverage that we can refactor these replacers with
confidence.