refactor(kubernetes): Push account filtering to artifact provider
All consumers of the artifact provider immediately filter the
results to a particular account; let's push this logic to the
artifact provider so they don't need to.

This removes the need for the checkArtifactAcrossAccount test,
which ensures that the versioned artifact converter, as it is
now replaced by a test on the artifact provider itself to ensure
that it only returns artifacts for a single account.