refactor(kubernetes): Use filters in JsonPath expressions
The Replacer class contains two compilcated strings representing a
JsonPath; in all cases these are almost identical. Let's replace
this logic with something a bit more clear.  Ideally we'd not need
to use JsonPath because we would have a strongly-typed object and
would be able to use that, but for now let's at least improve this
code (and open up some performance improvements to come).

First, findPath and replacePath are always of the exact same form;
we build a query to find certain nodes. The findPath grabs all of
them while the replacePath filters them to ones matching a supplied
artifact. In addition to duplicating logic, the syntax of filter
conditions is very hard to read.

The JsonPath library allows you to put a placeholder [?] to represent
a filter condition, then bind it to a supplied object representing
the condition. This is exactly what we want; we want the same query
in all cases but to bind to "match everything" in the find case and
to bind to "match this artifact" in the replace case.

Let's require the replacers to supply a path with a [?] placeholder,
and then to supply a filter to determine whether to match a particular
artifact. There are two replacers that also require filtering for the
find case, so we'll allow this as an optional field (but will default
to finding everything).

Also add a NonnullByDefault to the class which replaces the need for
some individual Nonnull annotations.