perf(kubernetes): Precompile find path for artifact replacers
The find path for artifact replacers is completely known a the time
the object is constructed. Let's pre-compile it so that we don't need
to do this compilation every time we look up artifats (which is every
time we return a manfiest from the /manifests endpoint).

We can't do the same with the replace query as it depends on the
input artifact. This isn't as much of an issue as we only replace
during a deploy rather than any time you request an existing manifest.
But let's slightly simplify the API by having the constructor handle
composing the find/replace queries.