refactor(kubernetes): Clean up filterArtifacts function
Rename the verbose filterKubernetesArtifactsByNamespaceAndAccount
to simply filterArtifacts.

To make the logic more clear split it out into a few predicates
and use those predicates in the filter logic.

Finally, require namespace to be non-null and push that requirement
outward. In one case there's something we don't know for sure is
non-null so convert it with nullToEmpty.