refactor(kubernetes): Simplify findGreatestUnusedVersion
There are a few simplifications to make here:
(1) We don't need to collect all the prior version to a sorted
array then grab the last element; we can just use max() to find the
maximum used version. This is slightly more efficient, O(N) vs
O(N log N) but really just makes the code simpler.
(2) We don't need to special-case the case of a version >= 1000;
the format string v%03d means pad to a *minimum* width of 3, and
handles numbers greater than 999 just fine.

Let's also add a couple of tests to the already good test coverage
to be sure.