refactor(kubernetes): Clean up exception handling
We don't need to log right before throwing an exception; a log
message is only useful when suppressing the exception, as it's going
to show up in the logs anyway.

When catching an IOException wrap it in an UncheckedIOException rather
than a RuntimeException.

Finally, don't catch a generic Exception, instead catch a RuntimeException.