refactor(kubernetes): Inline some variables
In both implementations of toArtifact we store a bunch of fields
from a KubernetesManifest to local variables in order to use them
a few lines later. This just adds more indirection and is confusing;
just read the fields we need from the KubernetesManifest.

Also getVersion should just take the KubernetesManifest itself;
it's just confusing to extract a bunch of fields into a long list
of arguments instead of just passing the manifest itself.