refactor(kubernetes): Minor cleanup of find/replace logic
The find function can return a Stream as it's only used in one
place and the collection it returns is immediately converted to
a Stream.

Minorly refactor the replace function as well.  The goal here is
just to make it more clear which functions have side effects by
not having them deeply nested in other functions.