perf(kubernetes): Only deserialize pod status
Rather than convert the entire manifest to a V1Pod, only convert the
status as that's all this function is using.

Ideally we'd only do a single conversion to the full appropriate type
when we get a value back from the cluster, and pass that around, but
as long as we're still doing these ad-hoc intercoversions, it's worth
avoiding doing more than we need.