refactor(kubernetes): Move name parsing logic to KubernetesCoordinates
A lot of the Kubernetes logic parses a kind-name String into a
Pair<KubernetesKind, String>. This leads to very unclear code as
we are always calling Pair#getLeft and Pair#getRight instead of using
more clearly-named functions. While Pairs might be reasonable for
short-term holding of data in functions or streams, they quickly
lead to complicated code when passed across function boundaries.

We already have a KubernetesCoordinates class that represents a kind
and a name (as well as a namespace). (And in most cases where we're
parsing a kind-name string we also have a namespace around too.)

Let's prepare to move clients to instead parse into KubernetesCoordinates.
This change does the following:
* Explicitly write out the auto-generated constructor and have it translate
null to empty for name/namespace. This allows us to guarantee that none of
the fields are null regardless of how the builder is used. We will explicitly
prohibit null KubernetesKind because we are always passing in a non-null kind (as
our kind parsing logic returns KubernetesKind.NONE instead of null).
* Add a method to the builder fullResourceName that takes a "kind name" String and
parses it to set both the kind and the name.  (The other auto-generated builder
methods will still be injected by Lombok, so you can still call .name() or .kind()
separately if you want.)
* Deprecate KubernetesManifest.fromFullResourceName; update it to call the method
we just wrote and wrap the result in a Pair.