refactor(kubernetes): Add some functions that accept KubernetesCoordinates
There are a lot of places where we pass around a separate kind/namespace/name
that would be much clearer if we could just directly pass KubernetesCoordinates
(which avoids unpacking a nice object into individual parameters).

Let's add a few of these functions where they will be useful, in all cases
augmenting a function that accepts these separately to become one that
accepts a KubernetesCoordinates. I've deprecated the old versions (and
will update usages in a later commit) and rewritten them to call the
new version.

I had to make a small change to topPod, which was only checking if
the pod was null (rather than null or empty); as we're going through
KubernetesCoordinates now we will never have a null name as that will
always be changed to empty string.