refactor(kubernetes): Refactor other locations to use new function
The various Provider classes as well as the on-demand caching agent
can be updated now to create a KubernetesCoordinates class instead of
parsing the name into a Pair.

This also simplifies later code in these functions, which can be updated
to pass the KubernetesCoordinates on to other functions rather than
extracting out the parts.

While making this change, update the catch blocks around the name parsing
to catch an IllegalArgumentException rather than a generic Exception.