refactor(kubernetes): Extract some common logic in provider code
When doing the last refactor I realized that a number of providers have
almost identical logic. Make a function in CacheUtils that takes an
account/namespace/name and returns the corresponding cache data item so
we don't need to keep duplicating this.

One small change (which I considered doing before and commented on) is
that we no longer filter relationships in the server group case. The
performance benefits here are not worth the additional complexity of
not re-using the code (particularly that we only do this filtering when
requesting a single server group and not when requesting them all).