test(kubernetes): Move tests on name parsing
Now that we've removed all uses of fromFullResourceName, let's
move those tests to the function that we're recommending (and what
the deprecated method itself calls). This just converts the exact
same tests to Java and moves them.

I'll leave this and the other deprecated methods around at least
until after we cut the release so there is some time for anyone
who might have a custom build to migrate.