perf(kubernetes): Request metrics and events in parallel
When deck requests a manifest, it wants the events and metrics as
well. This means that we make 3 sequential calls to the cluster,
each with a bit of latency, that ends up making the overall call
longer than needed.

Let's speed this up by sending the event and metrics requests
to an executor pool that runs them on a separate thread. This
allows us to only wait as long as the longest request instead
of the sum of all of the request times.