refactor(core): Add default implementations of deprecated methods
There are a few methods in the models that are deprecated in favor
of calling a different method. But because these don't have default
implementations, every implementation needs to override it which
adds more and more calls to these methods.

This commit adds default implementations that point to the recommended
method. I've removed the now-redundant override from the Kubernetes
implementations, but will let the maintainers of other providers remove
their overrides if they would like.