fix(kubernetes): Fix errorprone TypeParameterUnusedInFormals
[TypeParameterUnusedInFormals] Declaring a type parameter that is
only used in the return type is a misuse of generics: operations on
the type parameter are unchecked, it hides unsafe casts at invocations
of the method, and it interacts badly with method overload resolution.

This check is 100% correct here; the only advantage of using the generic
here is that we can hide the casting in getRequiredField. In fact, let's
just get rid of this method and force callers to do the casting; keeping
the method just to throw the exception seems more confusing than inlining.