fix(kubernetes): Fix error-prone issues
This fixes a few error-prone issues that were trivial to fix, and
were easier than suppressing the whole warning class.

[EmptyBlockTag] A block tag (@param, @return, @throws, @deprecated)
has an empty description. Block tags without descriptions don't add
much value for future readers of the code; consider removing the tag
entirely or adding a description.

In this case just delete the tag as the name is enough.

[UnusedNestedClass] This nested class is unused, and can be removed.

We were indeed not using this nested class.

[PublicConstructorForAbstractClass] Constructors of on an abstract class
can be declared protected as there is never a need for them to be public.

[UnnecessaryLambda] Returning a lambda from a helper method or saving
it in a constant is unnecessary; prefer to implement the functional
interface method directly and use a method reference instead.

Turns out we weren't using this lambda saved to a variable at all.

[UnusedMethod] Private method is never used.

error-prone doesn't seem to be able to detect that JUnit is using these
as a MethodSource for parametrized tests; just suppress the warnings here.

Finally, this isn't from error-prone, but there's a compiler warning that
KubernetesNamedAccountCredentials overrides equals but doesn't call the
super class. We don't want to do that here (because there's nothing in
the super class and because we only want some explicitly-included fields)
so add callSuper = false to stop the warning.