refactor(kubernetes): Enable compilation warnings
This commit enables compilation warnings for clouddriver-kubernetes.
It turns off a few categories that either generate too many false
positives for our code (with a detailed comment) or that have too
many actual issues we need to fix first.

The plan is to fix the categories of issues one-by-one and
re-enable them.