refactor(kubernetes): Suppress some unchecked warnings
KubernetesManifest is full of unchecked casts; let's just add
a comment to the class and suppress the warnings for any methods
that currently have unchecked casts.

As noted in the comment, we should really try to stop relying on
this class that extends HashMap and is full of these unsafe casts,
but for now we'll just suppress the warnings.

I have not enabled the Xlint:unchecked flag because there are still
a few other unchecked warnings that I'll leave for later.

Also let's make computedKind transient because the class implements
Serializable. It doesn't really matter because we don't use java
serialization, but is more correct because this is just a cached
value that is stored to avoid re-computing.