fix(*): Pin elasticsearch in clouddriver-web
The new gradle rich versions feature is requiring us to be explicit
about dependency conflicts (whereas before it would resolve things).

Given that clouddriver-elasticsearch and clouddriver-elasticsearch-aws
pin different versions of elasticsearch, including them both in
clouddriver-web is causing a build failure.

For now I've resolved this by just picking the version that was
resolved before explicitly, but we probably should try to get these
on the same version.