fix(kubernetes): Do a live lookup in the artifact provider
This is only used during deploys, so latency is not a big issue
here. This will fix race conditions with live manifest calls
enabled where operations later in a pipeline don't account for
operations that happened earlier in the pipeline.

For now I am just replacing the exact same logic with a live call,
including:
* Building the whole artifact (even though we only need a few fields)
* Sorting by creation timestamp instead of by version

This unblocks deleting the code that caches artifacts, which then
unblocks some futher simplifications.