refactor(kubernetes): Stop caching artifacts
Now that we always do a live lookup of artifacts, we can stop
caching them, as no code reads artifacts from the cache.

I've left the enum constant for artifact cache keys, and left
the core caching agent authoritative, so that existing entries
get evicted. We can remove this code after we've done a release
with it present.